iTPM_SU_PromotionType_Assistant.js keeping original content
Updated the iTPM_SU_PromotionType_Assistant.js file to ignore the changes from DEV 2 Branch. The original file content is correct and should not be replaced. The code should have an IF condition on subsidiary before trying to get the user subsidiary. Also, variable initialization and declaration should be in separate statements for easier troubleshooting.